+++ app/Plugin/Configurations/Minimal.php
@@ -13,6 +13,7 @@ defined( 'ABSPATH' ) || exit;
 use easyDirectoryListingForWordPress\Taxonomy;
 use ExternalFilesInMediaLibrary\ExternalFiles\Files;
 use ExternalFilesInMediaLibrary\ExternalFiles\Tools;
+use ExternalFilesInMediaLibrary\Plugin\Admin\Directory_Listing;
 use ExternalFilesInMediaLibrary\Plugin\Configuration_Base;
 use ExternalFilesInMediaLibrary\Services\Services;
 use WP_Term_Query;
@@ -87,25 +88,17 @@ class Minimal extends Configuration_Base {
 		}
 
 		// 3. Check the external sources.
-		$query = array(
-			'taxonomy'   => Taxonomy::get_instance()->get_name(),
-			'hide_empty' => false,
-			'count'      => false,
-		);
-		$terms = new WP_Term_Query( $query );
-		if ( is_array( $terms->terms ) ) { // @phpstan-ignore function.alreadyNarrowedType
-			foreach ( $terms->terms as $term ) {
-				// get the used service name.
-				$service_name = get_term_meta( $term->term_id, 'type', true );
+		foreach ( Directory_Listing::get_instance()->get_external_sources() as $term ) {
+			// get the used service name.
+			$service_name = get_term_meta( $term->term_id, 'type', true );
 
-				// bail if service name is empty.
-				if( empty( $service_name ) ) {
-					continue;
-				}
-
-				// add the name to the list.
-				$used_services[ $service_name ] = $service_name;
+			// bail if service name is empty.
+			if( empty( $service_name ) ) {
+				continue;
 			}
+
+			// add the name to the list.
+			$used_services[ $service_name ] = $service_name;
 		}
 
 		// 4. Get all services and disable the ones, which are not in our list.